editlyric work

+++ frontend/src/utils/usePostLike.js
@@ -1,27 +1,9 @@
 import { useContext, useState } from 'react'
 import actions from '../api.js'
 import { SongDataContext } from '../contexts/SongData'
-import TheContext from '../contexts/TheContext.js'
 
 export default function usePostLike() {
-  const { user } = useContext(TheContext)
-  const { homeFeedSongs, setHomeFeedSongs, allSongComments, setAllSongComments } =
-    useContext(SongDataContext)
-
-  const initialLikes = {
-    TYPE: 'SONG',
-    IS_LIKED: false,
-    USERS_LIKE_TO_DELETE: null,
-    TOTAL_LIKES: null,
-  }
-  const initialCommentLikes = {
-    TYPE: 'COMMENT',
-    IS_LIKED: false,
-    USERS_COMMENTLIKE_TO_DELETE: null,
-    TOTAL_COMMENT_LIKES: null,
-  }
-  const [likes, setLikes] = useState(initialLikes)
-  const [commentLikes, setCommentLikes] = useState(initialCommentLikes)
+  const { homeFeedSongs, setHomeFeedSongs } = useContext(SongDataContext)
 
   const addSongLike = (songId, setUsersLike) => {
     if (songId == null || setUsersLike == null) return
@@ -79,6 +61,7 @@ export default function usePostLike() {
   }
 
   const addCommentLike = async (commentId, songId, setUsersLike) => {
+    if (commentId == null || songId == null || setUsersLike == null) return
     await actions
       .addCommentLike({ commentId: commentId, date: new Date() })
       .then(res => {
@@ -103,7 +86,6 @@ export default function usePostLike() {
                   }
                 } else return comm
               })
-              console.log(each, 'lolololololollllllll')
               return { ...each, song: { ...each.song, song_comments: songComments } }
             } else return each
           }),
@@ -113,6 +95,8 @@ export default function usePostLike() {
   }
 
   const deleteCommentLike = async (commentId, songId, toDelete) => {
+    if (commentId == null || songId == null || toDelete == null) return
+
     await actions
       .deleteCommentLike({ commentId: commentId, likeToDelete: toDelete })
       .then(res => {
@@ -135,96 +119,12 @@ export default function usePostLike() {
                   }
                 } else return comm
               })
-              console.log(each, 'lolololololollllllll')
               return { ...each, song: { ...each.song, song_comments: songComments } }
             } else return each
           }),
         )
-        // updateCommentsArr(commentId, songId, commentLikes)
       })
       .catch(console.error)
-  }
-
-  const updateCommentsArr = (commentId, songId, newLikes) => {
-    let newCommentArr = allSongComments.map(song => {
-      if (song.songId === songId) {
-        let newCommentLikes = song.comments.map(comment => {
-          if (comment._id === commentId) {
-            return { ...comment, likes: newLikes }
-          } else {
-            return comment
-          }
-        })
-        return { ...song, comments: newCommentLikes }
-      } else {
-        return song
-      }
-    })
-    setAllSongComments(newCommentArr)
-  }
-
-  const handleInViewCommentLikes = async (commentId, songId) => {
-    if (commentId == null) return
-
-    let likeData = await allSongComments.filter(song => song.songId === songId)
-
-    const { liked, likeToDelete, likesArr } = await filterLikesArray(
-      likeData[0]?.comments,
-      '_id',
-      commentId,
-    )
-
-    setCommentLikes(prevCommentLikes => ({
-      ...prevCommentLikes,
-      IS_LIKED: liked,
-      TOTAL_COMMENT_LIKES: likesArr?.length,
-      USERS_COMMENTLIKE_TO_DELETE: likeToDelete,
-    }))
-  }
-
-  // const handleInViewLikes = async (songId) => {
-  //   if (songId == null) return
-  //   setLikes(initialLikes)
-
-  //   const { liked, likeToDelete, likesArr } = await filterLikesArray(allSongLikes, 'songId', songId)
-
-  //   setLikes(prevLikes => ({
-  //     ...prevLikes,
-  //     IS_LIKED: liked,
-  //     TOTAL_LIKES: likesArr?.length,
-  //     USERS_LIKE_TO_DELETE: likeToDelete
-  //   }))
-  // }
-
-  const filterLikesArray = async (array, key, id) => {
-    let liked = false
-    let likeToDelete = null
-    let likesArr = []
-
-    await array.filter(each => {
-      if (each[key] === id) {
-        likesArr = each.likes
-        likesArr?.filter(like => {
-          if (like.user === user?._id) {
-            liked = true
-            likeToDelete = like
-          }
-        })
-      }
-    })
-    return { liked, likeToDelete, likesArr }
-  }
-
-  const handlePostLike = (like, commentId, songId, songUserId) => {
-    if (like.TYPE === 'SONG') {
-      if (songUserId === user?._id) return
-      else if (like.IS_LIKED) return deleteSongLike(songId, like.USERS_LIKE_TO_DELETE)
-      else return addSongLike(songId)
-    } else {
-      if (like.IS_LIKED)
-        return deleteCommentLike(commentId, like.USERS_COMMENTLIKE_TO_DELETE, songId)
-      else return addCommentLike(commentId, songId)
-    }
   }
 
   return {
